+++ src/__tests__/posthog-core.js
@@ -406,7 +406,7 @@ describe('posthog core', () => {
                 $window_id: 'windowId',
                 $session_id: 'sessionId',
                 $is_identified: false,
-                $process_person: true,
+                $process_person_profile: true,
             })
         })
 
@@ -428,15 +428,15 @@ describe('posthog core', () => {
                 $session_id: 'sessionId',
                 $lib_custom_api_host: 'https://custom.posthog.com',
                 $is_identified: false,
-                $process_person: true,
+                $process_person_profile: true,
             })
         })
 
-        it("can't deny or blacklist $process_person", () => {
-            given('property_denylist', () => ['$process_person'])
-            given('property_blacklist', () => ['$process_person'])
+        it("can't deny or blacklist $process_person_profile", () => {
+            given('property_denylist', () => ['$process_person_profile'])
+            given('property_blacklist', () => ['$process_person_profile'])
 
-            expect(given.subject['$process_person']).toEqual(true)
+            expect(given.subject['$process_person_profile']).toEqual(true)
         })
 
         it('only adds token and distinct_id if event_name is $snapshot', () => {
@@ -467,7 +467,7 @@ describe('posthog core', () => {
             expect(given.subject).toEqual({
                 event_name: given.event_name,
                 token: 'testtoken',
-                $process_person: true,
+                $process_person_profile: true,
             })
         })
 
@@ -756,7 +756,7 @@ describe('posthog core', () => {
             given('overrides', () => ({
                 sessionRecording: {
                     afterDecideResponse: jest.fn(),
-                    startRecordingIfEnabled: jest.fn(),
+                    startIfEnabledOrStop: jest.fn(),
                 },
                 toolbar: {
                     maybeLoadToolbar: jest.fn(),