+++ src/types.ts
@@ -1,5 +1,6 @@
 import type { MaskInputOptions, SlimDOMOptions } from 'rrweb-snapshot'
 import { PostHog } from './posthog-core'
+import type { SegmentAnalytics } from './extensions/segment-integration'
 
 export type Property = any
 export type Properties = Record<string, Property>
@@ -142,16 +143,16 @@ export interface PostHogConfig {
     // Should only be used for testing. Could negatively impact performance.
     disable_compression: boolean
     bootstrap: BootstrapConfig
-    segment?: any
+    segment?: SegmentAnalytics
     __preview_send_client_session_params?: boolean
     disable_scroll_properties?: boolean
     // Let the pageview scroll stats use a custom css selector for the root element, e.g. `main`
     scroll_root_selector?: string | string[]
 
-    /** You can control whether events from PostHog-js have person processing enabled with the `process_person` config setting. There are three options:
-     * - `process_person: 'always'` _(default)_ - we will process persons data for all events
-     * - `process_person: 'never'` - we won't process persons for any event. This means that anonymous users will not be merged once they sign up or login, so you lose the ability to create funnels that track users from anonymous to identified. All events (including `$identify`) will be sent with `$process_person: False`.
-     * - `process_person: 'identified_only'` - we will only process persons when you call `posthog.identify`, `posthog.alias`, `posthog.setPersonProperties`, `posthog.group`, `posthog.setPersonPropertiesForFlags` or `posthog.setGroupPropertiesForFlags` Anonymous users won't get person profiles.
+    /** You can control whether events from PostHog-js have person processing enabled with the `person_profiles` config setting. There are three options:
+     * - `person_profiles: 'always'` _(default)_ - we will process persons data for all events
+     * - `person_profiles: 'never'` - we won't process persons for any event. This means that anonymous users will not be merged once they sign up or login, so you lose the ability to create funnels that track users from anonymous to identified. All events (including `$identify`) will be sent with `$process_person_profile: False`.
+     * - `person_profiles: 'identified_only'` - we will only process persons when you call `posthog.identify`, `posthog.alias`, `posthog.setPersonProperties`, `posthog.group`, `posthog.setPersonPropertiesForFlags` or `posthog.setGroupPropertiesForFlags` Anonymous users won't get person profiles.
      */
     person_profiles?: 'always' | 'never' | 'identified_only'
     /** @deprecated - use `person_profiles` instead  */